Update deploy scripts
