+++ deploy/1_nerif_token.migration.ts
@@ -1,0 +1,13 @@
+import { Deployer } from '@solarity/hardhat-migrate';
+
+import { NerifToken__factory, TokensVesting__factory } from '../generated-types/ethers';
+import { parseConfig } from './helpers/configParser';
+
+export = async (deployer: Deployer) => {
+  const config = parseConfig();
+
+  if (config.systemTokenData) {
+    await deployer.deploy(NerifToken__factory);
+    await deployer.deploy(TokensVesting__factory);
+  }
+};
