+++ deploy/3_operational_contracts.migration.ts
@@ -1,4 +1,4 @@
-import { Deployer, Reporter } from '@solarity/hardhat-migrate';
+import { Deployer } from '@solarity/hardhat-migrate';
 
 import {
   BillingManager__factory,
@@ -11,14 +11,13 @@ import {
   SignerStorage__factory,
 } from '../generated-types/ethers';
 import { parseConfig } from './helpers/configParser';
-import { ethers } from 'hardhat';
 
 export = async (deployer: Deployer) => {
   const config = parseConfig();
 
   const billingManagerImpl = await deployer.deploy(BillingManager__factory, {name: "BillingManagerImpl"});
   const registryImpl = await deployer.deploy(Registry__factory, {name: "RegistryImpl"});
-  const gatewayImpl = await deployer.deploy(Gateway__factory, {name: "GatewayImpl"});
+  await deployer.deploy(Gateway__factory, {name: "GatewayImpl"});
 
   const billingManagerProxy = await deployer.deploy(ERC1967Proxy__factory, [billingManagerImpl.address, "0x"], {name: "BillingManagerProxy"});
   const registryProxy = await deployer.deploy(ERC1967Proxy__factory, [registryImpl.address, "0x"], {name: "RegistryProxy"});
@@ -26,10 +25,7 @@ export = async (deployer: Deployer) => {
   await deployer.save(BillingManager__factory, billingManagerProxy.address);
   await deployer.save(Registry__factory, registryProxy.address);
 
-  const billingManager = await deployer.deployed(BillingManager__factory);
-  const registry = await deployer.deployed(Registry__factory);
-
-  const gatewayFactory = await deployer.deploy(GatewayFactory__factory);
+  await deployer.deploy(GatewayFactory__factory);
 
   let signerStorage: SignerStorage;
 
@@ -46,31 +42,4 @@ export = async (deployer: Deployer) => {
   } else {
     throw new Error("Undefined signer address");
   }
-
-  await billingManager.initialize(registry.address, signerStorage.address, {
-    depositAssetKey: config.operationalContractsInitParams.nativeDepositAssetData.nativeDepositAssetKey,
-    depositAssetData: {
-      tokenAddr: ethers.constants.AddressZero,
-      workflowExecutionDiscount: config.operationalContractsInitParams.nativeDepositAssetData.workflowExecutionDiscount,
-      isEnabled: config.operationalContractsInitParams.nativeDepositAssetData.isEnabled,
-      networkRewards: 0,
-      isPermitable: false,
-    }
-  });
-  await gatewayFactory.initialize(registry.address, gatewayImpl.address);
-  await registry.initialize(
-    config.isMainChain,
-    signerStorage.address,
-    gatewayFactory.address,
-    billingManager.address,
-    config.operationalContractsInitParams.maxWorkflowsPerAccount
-  );
-
-  Reporter.reportContracts(
-    ["BillingManager", billingManager.address],
-    ["Gateway Impl", gatewayImpl.address],
-    ["GatewayFactory", gatewayFactory.address],
-    ["Registry", registry.address],
-    ["SignerStorage", signerStorage.address],
-  );
 };
